🧹 Fix catalog gallery view
This commit will update the Hyku submodule which will bring in the fix
for the catalog gallery view.  Also, it removes the
truncate_and_iconify_auto_link method because it got contributes back to
Hyku.  We are also removing all the underline text decoration from the
a tags but keeping the underline for the hover and active effect.

Ref:
- https://github.com/notch8/palni_palci_knapsack/issues/174