🧹 Set Bulkrax default work
This commit will set the default work for Bulkrax explicitly.  Prior, it
was falling back to Cdl, because by default it looks at the first of the
Hyrax.config.curation_concerns.  Setting it to GenericWorkResource will
make sure if any imports with no explicitly set model that the works
don't create Cdl works instead, which is a very specific work type.