⬆️ Submodule update to disable batch_upload
This Hyku bump prevents admin users from enabling the batch_upload
feature in settings. This addresses the issue that batch uploads and
Valkyrized works are incompatible. The current Batch Add uses a fake
Active Fedora form with fixed terms that ignore the selected work type
causing Valkyrie validations to fail (especially on 6.2).

It was determined that fixing this issue is low-priority compared to
general Bulkrax improvements. To ensure users are not enabling broken
features, batch_upload is disabled until further decisions are made.

Ref:
- notch8/palni_palci_knapsack#479