🐛 Controlled vocabulary fields are dropdowns
In v6.2 when flexible=false, the controlled vocabulary fields rendered a
text field in the new/edit forms rather than a dropdown. This allowed
the user to enter data that did not match a predefined mapping
(e.g. `term: "In Copyright"`) defined in the hyku/config/authorities yml
files. Controlled vocabulary maps the term from the form dropdown to the
id that is persisted
(e.g. `id: http://rightsstatements.org/vocab/InC/1.0/`).

Ref:
- https://github.com/notch8/palni_palci_knapsack/issues/495